Comment add in admin panel

+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Role;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -27,8 +28,12 @@ class UserController extends Controller
      */
     public function create()
     {
-        $data['roles'] = Role::orderBy('name')->get();
-        return view('admin.user.create', $data);
+        if(Gate::allows('admin',auth()->user())) {
+            $data['roles'] = Role::orderBy('name')->get();
+            return view('admin.user.create', $data);
+        }
+        session()->flash('success','You are not authorized to perform this operation.');
+        return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -39,20 +44,24 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required',
-            'email' => 'required|email|unique:users',
-            'password' => 'required|confirmed',
-            'role_id' => 'required',
-            'status' => 'required',
-        ]);
+        if(Gate::allows('admin',auth()->user())) {
+            $request->validate([
+                'name' => 'required',
+                'email' => 'required|email|unique:users',
+                'password' => 'required|confirmed',
+                'role_id' => 'required',
+                'status' => 'required',
+            ]);
 
-        $data = $request->except(['_token', 'password']);
-        $data['password'] = bcrypt($request->password);
+            $data = $request->except(['_token', 'password']);
+            $data['password'] = bcrypt($request->password);
 
-        User::create($data);
-        session()->flash('success', 'User Create Successfully');
-        return redirect()->route('user.index');
+            User::create($data);
+            session()->flash('success', 'User Create Successfully');
+            return redirect()->route('user.index');
+        }
+      session()->flash('success','You are not authorized to perform this operation.');
+      return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -74,9 +83,13 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        if(Gate::allows('admin',auth()->user())) {
         $data['roles'] = Role::orderBy('name')->get();
         $data['user'] = User::findOrFail($id);
         return view('admin.user.edit', $data);
+        }
+        session()->flash('success','You are not authorized to perform this operation.');
+        return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -88,25 +101,29 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'name' => 'required',
-            'email' => 'required|email',
-            'role_id' => 'required',
-            'status' => 'required',
-        ]);
+        if(Gate::allows('admin',auth()->user())) {
+            $request->validate([
+                'name' => 'required',
+                'email' => 'required|email',
+                'role_id' => 'required',
+                'status' => 'required',
+            ]);
 
-        if ($request->password != null){
-            $data['password']=bcrypt($request->password);
+            if ($request->password != null){
+                $data['password']=bcrypt($request->password);
+            }
+            $data['name'] = $request->name;
+            $data['email'] = $request->email;
+            $data['role_id'] = $request->role_id;
+            $data['status'] = $request->status;
+
+            $user = User::findOrFail($id);
+            $user->update($data);
+            session()->flash('success', 'User Update Successfully');
+            return redirect()->route('user.index');
         }
-        $data['name'] = $request->name;
-        $data['email'] = $request->email;
-        $data['role_id'] = $request->role_id;
-        $data['status'] = $request->status;
-
-        $user = User::findOrFail($id);
-        $user->update($data);
-        session()->flash('success', 'User Update Successfully');
-        return redirect()->route('user.index');
+        session()->flash('success','You are not authorized to perform this operation.');
+        return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -117,9 +134,13 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user = User::findOrFail($id);
-        $user->destroy($user->id);
-        session()->flash('success', 'User Delated Successfully');
-        return redirect()->back();
+        if(Gate::allows('admin',auth()->user())) {
+            $user = User::findOrFail($id);
+            $user->destroy($user->id);
+            session()->flash('success', 'User Delated Successfully');
+            return redirect()->back();
+        }
+        session()->flash('success','You are not authorized to perform this operation.');
+        return redirect()->route('admin.dashboard');
     }
 }
